split file: renderer.ts parser.ts
